requirements.txt
Pillow
pytesseract
tqdm
reportlab